Everything is not entirely funcioning, still some test codes
