+++ main.py
@@ -1,5 +1,5 @@
 import matplotlib.pyplot as plt
-from mpl_toolkits import mplot3d
+#from mpl_toolkits import mplot3d
 from mpl_toolkits.mplot3d import axes3d
 import numpy as np
 import math
